+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#357](https://github.com/babylonlabs-io/vigilante/pull/357) imp: handle modules genesis changes
+
 ## v0.23.6
 
 ### Improvements